update code Ass 4 P2

+++ Assignment_4_2/inc/gemtek_lib.h
@@ -5,14 +5,16 @@
 #include <stdbool.h>
 #include <math.h>
 #include <string.h>
+#include <limits.h>
+#include <stdio_ext.h>
 
 void get_input_integer(int* integer_number);
 
 void print_square_cube_integer(int integer_number);
 
-int calculate_square_integer(int integer_number);
+double calculate_square_integer(int integer_number);
 
-int calculate_cube_integer(int integer_number);
+double calculate_cube_integer(int integer_number);
 
 int is_fibonacci(int integer_number);
